Splited PlotGroup and Plot; added some methods
